Feature/ros2 carma node handle (#83)
This PR adds a CarmaLifecycleNode base class which will serve as the base class for all ROS2 nodes in carma. This is effectively a replacement for the CARMANodeHandle from ros1 and provides similar exception handling behavior. The CarmaLifecycleNode follows the structure of a rclcpp::LifecycleNode allowing for management of the node lifecycle by external components. If an error occurs in this node at runtime it will be caught and logged. In addition, an attempt will be made to publish a system_alert message containing the exception. The lifecycle behavior in ROS2 is documented here https://design.ros2.org/articles/node_lifecycle.html
However, one critical oversight of the current ros implementation is that exception handling is not supported it primary states. This base class resolves this problem.

Extending classes are prevented from interfering in default behavior using the template pattern where the on_activate type methods are replaced with handle_on_activate.

The primary state exception behavior is implemented through overrides to all callback registration methods (subscribers, timers, services, and parameter callbacks). You will notice there is no system_alert subscriber only a publisher. This is intentional as the lifecycle of these nodes is managed externally not internally.

To make management of carma lifecycle nodes easier, I have create a ros2_lifecycle_manager which can be provided with a set of nodes that it will manage. It is effectively a signal multicaster. The user calls configure() for example and the manager will send that lifecycle signal to all managed nodes. This manager will be used in other PRs to enable system management.

The cmake in this leverages carma_cmake_common which has been added to carma-msgs

There is still a little bit of cleanup for me to do on this, but its fairly minimal.