Feature/helper methods for parameters (#87)
This PR adds an update_params method to the CarmaLifecycleNode which provides a mechanism for easily updating multiple parameters of the same type by reference or function when the parameter update callback is triggered. The parameter callback method in ROS doesn't allow for easy identification of which parameters have changed without tedious loops and type checks. This function is meant to reduce that load by accepting a map of parameter name to variable reference. It will then read the updated parameter as the appropriate type and assign it.