Update ros2_lifecycle_manager.cpp (#131)
Description
This PR increases the timeout for get state calls in the ros2_lifecycle_manager to improve startup stability for plugins. This doesn't seem to be a 100% fix, and it might be better to make this a configurable parameter. But this should help in the short term.

Related Issue
usdot-fhwa-stol/carma-platform#1899