Fix/occlusion fix (#198)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description

<!--- Describe your changes in detail -->
**Y value fix:**
Previously we discovered that the Y value of CARLA objects are wrong. 
We incorrectly assumed that the hitpoints report absolute values.
hitpoints from sensor is probably correct as they are in their own frame
(small values, smaller the distance).
However, the object's locations and sensor's locations all report in
negative Y value. Therefore, it is the best to do all the calculations
as is and only change the Y value before reporting the detected objects
to the caller of the sensor.

Also for complete occlusion-fix, this is accompanied by this PR:
https://github.com/usdot-fhwa-stol/carma-carla-integration/pull/54/files

More Data:
Data about actors in the scene where CAR is actually lower Y value in
correct map, but the data returned from CARLA is inverted:

> [DetectedObject(objectId=207, type='CAR', position=array([ 237.710327,
-169.307953, 2.32106090e-01]), DetectedObject(objectId=220,
type='TRUCK', position=array([ 239.819992, -173.389999, 1.90679163e-01])

 
Sensor data returned from the sensor has also inverted Y:

> pre_association:  [ 2.125664   -3.06154299  0.33665588]
> pre_association:  [ 2.26882815 -3.03080082 -1.09483194]
> pre_association:  [ 4.30117178 -2.98634768 -0.83236283]
> pre_association:  [ 1.90921867 -3.21779299 -0.35920775]
> pre_association:  [ 1.89583981 -3.1952734  -0.82916898]

I am not sure how the internal CARLA logic works, but the Y in lidar
hitpoint is also inverted as shown above which should report positive Y
values.

## Related GitHub Issue
Y value fix:
Fix for PR: https://github.com/usdot-fhwa-stol/carma-utils/pull/191

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key
Closes: https://github.com/usdot-fhwa-stol/carma-utils/issues/190

<!-- e.g. CAR-123 -->
Closes [CDAR-534](https://usdot-carma.atlassian.net/browse/CDAR-534)
## Motivation and Context

## How Has This Been Tested?
CDASIM integration tested, platform reports y values in positive
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [X] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [X] I have added any new packages to the sonar-scanner.properties file
- [X] My change requires a change to the documentation.
- [X] I have updated the documentation accordingly.
- [X] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [X] I have added tests to cover my changes.
- [X] All new and existing tests passed.



[CDAR-534]:
https://usdot-carma.atlassian.net/browse/CDAR-534?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ