Modernize `ros2_lifecycle_manager` CMake configuration (#219)
# PR Details
## Description

This PR modernizes the `ros2_lifecycle_manager`'s CMake configuration.
It also resolves an issue preventing the package's tests from building
in Humble. It seems like one of the built test executables was not
linking to `rclcpp`, causing the build to fail.

The package's CMake configuration replaces the legacy variable-style
with a modern target-based approach. It also merges the
`test/CMakeLists.txt` file with the top-level one because `ament_auto`
does not seem to work well with hierarchical CMake structures.

## Related GitHub Issue

## Related Jira Key

Closes [CF-803](https://usdot-carma.atlassian.net/browse/CF-803)
[CF-793](https://usdot-carma.atlassian.net/browse/CF-793)

## Motivation and Context

Need Humble support for C1T project

## How Has This Been Tested?

Built and run unit tests

## Types of changes

- [x] Defect fix (non-breaking change that fixes an issue)

## Checklist:

- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [x] All new and existing tests passed.


[CF-803]:
https://usdot-carma.atlassian.net/browse/CF-803?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[CF-793]:
https://usdot-carma.atlassian.net/browse/CF-793?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ