Add Humble forward compatibility support for `rclcpp_lifecycle::LifecyclePublisher`s (#220)
# PR Details
## Description

This PR adds forward compatibility support for ROS Humble for the
`rclcpp_lifecycle::LifecyclePublisher`. The `LifecyclePublisher` class
uses a different base class in Humble than it does in Foxy. The base
class from Foxy (`rclcpp_lifecycle::LifecyclePublisherInterface`) was
removed between Foxy and Humble.

We still need to support both ROS releases for now, so we need a shim
that can switch conditionally between the two APIs.

## Related GitHub Issue

## Related Jira Key

Closes [CF-811](https://usdot-carma.atlassian.net/browse/CF-811)
[CF-809](https://usdot-carma.atlassian.net/browse/CF-809)

## Motivation and Context

Needed for C1T development

## How Has This Been Tested?

Compiled with both Foxy and Humble. Humble has build errors, but those
are unrelated to this change.

## Types of changes

- [x] New feature (non-breaking change that adds functionality)

## Checklist:

- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.


[CF-811]:
https://usdot-carma.atlassian.net/browse/CF-811?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[CF-809]:
https://usdot-carma.atlassian.net/browse/CF-809?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ