Fix quintic polynomial equation solver (#249)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description

<!--- Describe your changes in detail -->
After we solve the equation, sometimes there was a slight deviation
along the trajectory when we use back the equation to get speed and
downtrack, and I was wondering why. Here is what I changed:

- Changed __uint64_t to double for time parameters to avoid losing
precision as we use it with small double variables (remaining_time left
to plan around 15.0 or so seconds).
- Uses a solver function instead of manually inverting and getting the
answer. This method avoids dividing by zero.

https://eigen.tuxfamily.org/dox/classEigen_1_1ColPivHouseholderQR.html#ad4825c3d43dffdf0c883de09ba891646
Seems more robust and I can actually see the if we use the coefficients
the result is what we expect.

The original PR for this change is interestingly deleted or not
accessible. So I am not sure why the uint64 was chosen except maybe it
was trying to force positive number because it is a time.

## Related GitHub Issue
https://github.com/usdot-fhwa-stol/carma-platform/issues/2345
Will improve above but not necessarily fix all the way
<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key
NA
<!-- e.g. CAR-123 -->

## Motivation and Context
Improve Yield
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Sim PC
Resulting Coefficients actually result in matching output with online
solver with given t (or x)
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [X] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [X] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [X] I have updated the documentation accordingly.
- [X] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [X] I have added tests to cover my changes.
- [X] All new and existing tests passed.